+++ include/app/watch.h
@@ -1,0 +1,6 @@
+#ifndef KEYVIBE_WATCH_H
+#define KEYVIBE_WATCH_H
+
+int start_config_watcher(const char *path);
+
+#endif // KEYVIBE_WATCH_H
