+++ include/common/utils.h
@@ -7,7 +7,8 @@
 // Returns 1 on success, 0 on truncation/error
 int safe_snprintf(char *dst, size_t dst_sz, const char *fmt, ...);
 
-// fprintf to stderr with printf-style formatting. Returns 1 for convenient `return errorf(...)` usage.
+// fprintf to stderr with printf-style formatting. Returns 1 for convenient
+// `return errorf(...)` usage.
 int errorf(const char *fmt, ...);
 
 // Common runtime helpers
@@ -24,4 +25,3 @@ void write_runtime_keyboard_mute_file(int mute);
 void write_runtime_mouse_mute_file(int mute);
 
 #endif // KEYVIBE_UTILS_H
-
