+++ src/input.c
@@ -1,4 +1,4 @@
-#include "utils.h"
+#include "common/utils.h"
 #include <errno.h>
 #include <stdbool.h>
 #include <stdint.h>
